Extract dialogs and pane controls into file_manager subpackage
Phase 4b of decomposing file_manager_window. Move the standalone
dialogs and pane-level chrome out:

- file_manager/progress_dialog.py — SFTPProgressDialog plus the
  _HAS_ALERT_DIALOG / _PROGRESS_DIALOG_BASE shim that selects
  between Adw.AlertDialog and the deprecated Adw.MessageDialog.
- file_manager/pane_controls.py — PathEntry, PaneControls,
  PaneToolbar (path bar, nav buttons, toolbar with view-mode and
  show-hidden toggles).
- file_manager/properties_dialog.py — Nautilus-style
  PropertiesDialog. Uses TYPE_CHECKING for forward refs to
  FileEntry and AsyncSFTPManager so it doesn't pull the
  file_manager_window mega-module at import time.

file_manager_window.py re-imports the public names from
sshpilot.file_manager, so existing callers and the manual SFTP
scripts keep working unchanged. Net: file_manager_window.py
8491 -> 7360 lines.

AsyncSFTPManager, FilePane, and FileManagerWindow still live in
file_manager_window.py and will move in phase 4c.

Refs #11
