+++ sshpilot/file_manager_window.py
@@ -55,7 +55,14 @@ from .platform_utils import is_flatpak, is_macos
 from .text_editor import RemoteFileEditorWindow
 from .file_manager import (
     DOCS_JSON,
+    PaneControls,
+    PaneToolbar,
+    PathEntry,
+    PropertiesDialog,
+    SFTPProgressDialog,
     TransferCancelledException,
+    _HAS_ALERT_DIALOG,
+    _PROGRESS_DIALOG_BASE,
     _ensure_cfg_dir,
     _get_docs_json_path,
     _grant_persistent_access,
@@ -82,515 +89,6 @@ import logging
 logger = logging.getLogger(__name__)
 
 
-
-# ---------------------------------------------------------------------------
-# Utility data structures
-
-
-# Prefer Adw.AlertDialog (libadwaita ≥ 1.5, May 2024) — Adw.MessageDialog is
-# deprecated since 1.6. Fall back to MessageDialog on older systems so the
-# file manager still works there. The two classes differ in:
-#   * constructor: AlertDialog uses ``heading=``, MessageDialog uses ``title=``
-#   * setter: ``set_heading`` vs ``set_title``
-#   * present: AlertDialog takes a parent widget; MessageDialog uses
-#     set_transient_for + set_modal then present()
-# All three differences are bridged below.
-_HAS_ALERT_DIALOG = hasattr(Adw, "AlertDialog")
-_PROGRESS_DIALOG_BASE = Adw.AlertDialog if _HAS_ALERT_DIALOG else Adw.MessageDialog
-
-
-class SFTPProgressDialog(_PROGRESS_DIALOG_BASE):
-    """Modern GNOME HIG-compliant SFTP file transfer progress dialog.
-
-    Subclasses ``Adw.AlertDialog`` when available (libadwaita ≥ 1.5) and
-    falls back to the deprecated ``Adw.MessageDialog`` on older systems.
-
-    UI is driven from a fixed-cadence ``GLib.timeout`` (the canonical GTK
-    ProgressBar pattern). The worker's per-chunk callbacks only mutate a few
-    state fields and push samples into a sliding window — they do not touch
-    widgets. The render tick reads that state and updates labels + bar at
-    ~4 Hz, which keeps the UI responsive without burning cycles on transfers
-    that emit thousands of callbacks per second.
-    """
-
-    _LABEL_WIDTH_CHARS = 48
-    _RENDER_INTERVAL_MS = 250
-    _SPEED_WINDOW_SECONDS = 5.0
-
-    def __init__(self, parent=None, operation_type="transfer"):
-        title = "Downloading Files" if operation_type == "download" else "Uploading Files"
-
-        # Different constructor kwargs for the two base classes.
-        if _HAS_ALERT_DIALOG:
-            super().__init__(
-                heading=title,
-                body="Transferring files…",
-                default_response="cancel",
-            )
-        else:
-            super().__init__(
-                title=title,
-                body="Transferring files…",
-                default_response="cancel",
-            )
-            # MessageDialog is a Gtk.Window — old API: set transient + modal.
-            if parent is not None:
-                try:
-                    self.set_transient_for(parent)
-                except Exception:
-                    pass
-            try:
-                self.set_modal(True)
-            except Exception:
-                pass
-
-        # Transfer state
-        self.is_cancelled = False
-        self.current_file = ""
-        self.files_completed = 0
-        self.total_files = 0
-        self.operation_type = operation_type
-        self._current_future = None
-        self._futures = []  # Track all futures for multi-file operations
-        self._completion_shown = False
-        self._failed_files = []
-
-        # Tracks whether the dialog is currently presented. Used to make
-        # close() idempotent so that the shutdown cleanup path (which calls
-        # close again after the user has dismissed the dialog) doesn't trip
-        # the "trying to close a dialog that's not presented" Adwaita-CRITICAL.
-        self._closed = False
-
-        # Bytes-driven speed/ETA state (populated by progress-bytes signal).
-        # Always read/written on the main thread.
-        self._transferred_bytes = 0
-        self._total_bytes = 0
-        # Sliding window of (monotonic_time, transferred_bytes) samples used
-        # to compute the *recent* throughput rather than the lifetime average.
-        self._byte_samples: "collections.deque[tuple[float, int]]" = collections.deque()
-
-        # Latest fraction/message/file pushed by ``update_progress``. The
-        # render tick reads these — neither the worker callback nor the
-        # signal handler touches widgets directly.
-        self._latest_fraction: Optional[float] = None
-        self._latest_message: Optional[str] = None
-        self._latest_file: Optional[str] = None
-
-        self._build_ui()
-
-        # Start the render loop. Returning False from _render_tick removes the
-        # source; we also clear the id on dialog close as a belt-and-braces.
-        self._render_timeout_id: Optional[int] = GLib.timeout_add(
-            self._RENDER_INTERVAL_MS, self._render_tick
-        )
-        
-    def _build_ui(self):
-        """Build the modern GNOME HIG-compliant UI.
-
-        We deliberately do NOT add an Adw response button for Cancel. Single
-        responses in Adw.AlertDialog / Adw.MessageDialog span the entire
-        footer width — too prominent for a progress dialog. Instead, Cancel
-        is a regular Gtk.Button placed at the bottom of the body content,
-        right-aligned. Esc-to-cancel is preserved via the ``closed`` signal
-        (Adw.Dialog fires that for any close path, including Esc).
-        """
-
-        # No add_response(); no set_default_response(); no set_close_response().
-        # We listen on ``closed`` to catch Esc / parent dismiss / etc.
-        self.connect("closed", self._on_dialog_closed)
-
-        # Create progress content area
-        progress_box = Gtk.Box(
-            orientation=Gtk.Orientation.VERTICAL,
-            spacing=12,
-            margin_top=12,
-            margin_bottom=12
-        )
-        
-        def _configure_progress_label(label: Gtk.Label) -> None:
-            label.set_ellipsize(Pango.EllipsizeMode.END)
-            label.set_justify(Gtk.Justification.CENTER)
-            label.set_width_chars(self._LABEL_WIDTH_CHARS)
-            label.set_max_width_chars(self._LABEL_WIDTH_CHARS)
-
-        def _configure_path_label(label: Gtk.Label) -> None:
-            # Middle-ellipsis keeps both the head (drive/scheme/leading dir)
-            # and tail (filename) visible — best fit for paths.
-            label.set_ellipsize(Pango.EllipsizeMode.MIDDLE)
-            label.set_halign(Gtk.Align.START)
-            label.set_xalign(0.0)
-            label.set_width_chars(self._LABEL_WIDTH_CHARS)
-            label.set_max_width_chars(self._LABEL_WIDTH_CHARS)
-            label.add_css_class("caption")
-            label.add_css_class("dim-label")
-
-        # Current file label (primary info)
-        self.file_label = Gtk.Label()
-        self.file_label.set_text("—")
-        _configure_progress_label(self.file_label)
-        progress_box.append(self.file_label)
-
-        # Status label for detailed progress messages
-        self.status_label = Gtk.Label()
-        self.status_label.set_text("Preparing transfer…")
-        _configure_progress_label(self.status_label)
-        progress_box.append(self.status_label)
-
-        # Source / destination paths. Hidden until set_paths() is called so
-        # we don't leave two empty "From:" / "To:" lines for callers that
-        # didn't supply them.
-        self.source_label = Gtk.Label()
-        _configure_path_label(self.source_label)
-        self.source_label.set_visible(False)
-        progress_box.append(self.source_label)
-
-        self.dest_label = Gtk.Label()
-        _configure_path_label(self.dest_label)
-        self.dest_label.set_visible(False)
-        progress_box.append(self.dest_label)
-
-        # Main progress bar
-        self.progress_bar = Gtk.ProgressBar()
-        self.progress_bar.set_show_text(True)
-        self.progress_bar.set_text("0%")
-        progress_box.append(self.progress_bar)
-        
-        # Transfer details
-        details_box = Gtk.Box(
-            orientation=Gtk.Orientation.VERTICAL,
-            spacing=6
-        )
-        progress_box.append(details_box)
-        
-        # Speed and time info
-        info_box = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL)
-        details_box.append(info_box)
-        
-        self.speed_label = Gtk.Label()
-        self.speed_label.set_text("—")
-        self.speed_label.set_halign(Gtk.Align.START)
-        self.speed_label.add_css_class("caption")
-        info_box.append(self.speed_label)
-        
-        # Spacer
-        spacer = Gtk.Box()
-        spacer.set_hexpand(True)
-        info_box.append(spacer)
-        
-        self.time_label = Gtk.Label()
-        self.time_label.set_text("—")
-        self.time_label.set_halign(Gtk.Align.END)
-        self.time_label.add_css_class("caption")
-        info_box.append(self.time_label)
-        
-        # File counter
-        self.counter_label = Gtk.Label()
-        self.counter_label.set_text("0 of 0 files")
-        self.counter_label.set_halign(Gtk.Align.CENTER)
-        self.counter_label.add_css_class("caption")
-        details_box.append(self.counter_label)
-
-        # Cancel / Done action button. Placed in the body (not as an Adw
-        # response) so it doesn't get the full-width single-response styling.
-        action_row = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL)
-        action_row.set_halign(Gtk.Align.END)
-        action_row.set_margin_top(6)
-        self.action_button = Gtk.Button(label="Cancel")
-        self.action_button.add_css_class("pill")
-        self.action_button.connect("clicked", self._on_action_button_clicked)
-        action_row.append(self.action_button)
-        progress_box.append(action_row)
-
-        # Set the progress content as extra child
-        self.set_extra_child(progress_box)
-    
-    def set_operation_details(self, total_files, filename=None):
-        """Set the operation details"""
-        # Only update total_files if it's larger (for adding more files to existing dialog)
-        if total_files > self.total_files:
-            self.total_files = total_files
-            self.counter_label.set_text(f"{self.files_completed} of {total_files} files")
-
-        if filename:
-            self.current_file = filename
-            self.file_label.set_text(filename)
-
-    def set_paths(self, source: Optional[str] = None,
-                  destination: Optional[str] = None) -> None:
-        """Display the source and destination paths.
-
-        Each label is middle-ellipsized so long paths stay readable; the
-        full untruncated path is exposed as a tooltip on hover.
-        """
-        if source:
-            self.source_label.set_text(f"From: {source}")
-            self.source_label.set_tooltip_text(source)
-            self.source_label.set_visible(True)
-        if destination:
-            self.dest_label.set_text(f"To: {destination}")
-            self.dest_label.set_tooltip_text(destination)
-            self.dest_label.set_visible(True)
-    
-    def _on_action_button_clicked(self, _button: Gtk.Button) -> None:
-        """Single click handler for the Cancel/Done button.
-
-        Behaviour depends on which state the dialog is in: before completion
-        the button cancels all tracked futures; after completion it just
-        closes the dialog.
-        """
-        if self._completion_shown:
-            # Already in "Done" mode — just close.
-            self._stop_render_timer()
-            self.close()
-            return
-        # Active-transfer mode — cancel everything.
-        self._cancel_active_transfers()
-        self._stop_render_timer()
-        self.close()
-
-    def _on_dialog_closed(self, _dialog) -> None:
-        """Esc / parent-dismiss / WM-close all funnel through here.
-
-        If the dialog is dismissed while a transfer is still running (no
-        completion shown yet), treat it as a cancel so we don't leave the
-        worker uploading bytes in the background.
-        """
-        if not self._completion_shown and not self.is_cancelled:
-            self._cancel_active_transfers()
-        self._stop_render_timer()
-        # Mark as closed so a subsequent .close() (e.g. from the file
-        # manager's shutdown cleanup) becomes a silent no-op.
-        self._closed = True
-
-    def close(self) -> bool:  # type: ignore[override]
-        """Idempotent close — safe to call after the dialog is already gone.
-
-        Adw.Dialog.close() asserts that the dialog is currently presented and
-        emits an Adwaita-CRITICAL otherwise. The file manager calls close()
-        from several places (action button, completion, shutdown cleanup),
-        and timing means we sometimes hit a dialog that has already been
-        dismissed. Bail early in that case instead of letting the assert
-        fire.
-        """
-        if getattr(self, "_closed", False):
-            return False
-        self._closed = True
-        try:
-            return super().close()
-        except Exception:
-            return False
-
-    def _cancel_active_transfers(self) -> None:
-        """Flip the cancel flag and call .cancel() on every tracked future."""
-        self.is_cancelled = True
-        if self._current_future and hasattr(self._current_future, 'cancel'):
-            try:
-                self._current_future.cancel()
-            except Exception:
-                pass
-        for future in self._futures:
-            if future and hasattr(future, 'cancel') and not future.done():
-                try:
-                    future.cancel()
-                except Exception:
-                    pass
-
-    def _stop_render_timer(self) -> None:
-        """Remove the render GLib.timeout source if it's still scheduled."""
-        if self._render_timeout_id is not None:
-            try:
-                GLib.source_remove(self._render_timeout_id)
-            except Exception:
-                pass
-            self._render_timeout_id = None
-    
-    def update_progress(self, fraction, message=None, current_file=None):
-        """Stash the latest fraction/message/file. Render tick paints it.
-
-        Cheap: no widget access. Safe to call from any thread (we hop to the
-        main loop). The caller is expected to have already computed the
-        overall-progress fraction for multi-file batches — this dialog does
-        not re-apply ``(files_completed + fraction) / total_files``.
-        """
-        GLib.idle_add(self._update_progress_state, fraction, message, current_file)
-
-    def _update_progress_state(self, fraction, message, current_file):
-        if fraction is not None:
-            self._latest_fraction = fraction
-        if message:
-            self._latest_message = message
-        if current_file:
-            self.current_file = current_file
-            self._latest_file = current_file
-        return False
-
-    def on_bytes(self, transferred: int, total: int) -> None:
-        """Receive raw byte counts from the manager (main thread)."""
-        try:
-            transferred_int = int(transferred or 0)
-            total_int = int(total or 0)
-        except (TypeError, ValueError):
-            return
-        self._transferred_bytes = transferred_int
-        if total_int > 0:
-            self._total_bytes = total_int
-        now = time.monotonic()
-        self._byte_samples.append((now, transferred_int))
-        cutoff = now - self._SPEED_WINDOW_SECONDS
-        while self._byte_samples and self._byte_samples[0][0] < cutoff:
-            self._byte_samples.popleft()
-
-    def _render_tick(self) -> bool:
-        """Repaint speed/ETA/progress at fixed cadence. Returns False to stop."""
-        # If the dialog has been dismissed, drop the timer.
-        try:
-            visible = self.get_visible()
-        except Exception:
-            visible = False
-        if not visible or self._completion_shown:
-            self._render_timeout_id = None
-            return False
-
-        # Status + current-file text.
-        if self._latest_message:
-            try:
-                self.status_label.set_text(self._latest_message)
-            except (AttributeError, RuntimeError):
-                pass
-        if self._latest_file:
-            try:
-                self.file_label.set_text(self._latest_file)
-            except (AttributeError, RuntimeError):
-                pass
-
-        # Progress bar: pulse when we don't yet know what we're transferring,
-        # otherwise show the fraction the caller computed.
-        try:
-            if self._total_bytes <= 0 and self._latest_fraction in (None, 0.0):
-                self.progress_bar.pulse()
-                self.progress_bar.set_text("")
-            elif self._latest_fraction is not None:
-                fraction = max(0.0, min(1.0, float(self._latest_fraction)))
-                self.progress_bar.set_fraction(fraction)
-                self.progress_bar.set_text(f"{int(fraction * 100)}%")
-        except (AttributeError, RuntimeError):
-            pass
-
-        # Speed (sliding-window) + ETA from real byte counts.
-        speed_text = None
-        eta_text = None
-        if len(self._byte_samples) >= 2:
-            t0, b0 = self._byte_samples[0]
-            t1, b1 = self._byte_samples[-1]
-            dt = t1 - t0
-            db = b1 - b0
-            if dt > 0 and db > 0:
-                bps = db / dt
-                speed_text = self._format_speed(bps)
-                if self._total_bytes > 0 and bps > 0:
-                    remaining = max(0, self._total_bytes - self._transferred_bytes)
-                    eta_text = (
-                        "Almost done…" if remaining == 0
-                        else self._format_time(remaining / bps)
-                    )
-        try:
-            self.speed_label.set_text(speed_text or "—")
-            self.time_label.set_text(eta_text or "—")
-        except (AttributeError, RuntimeError):
-            pass
-
-        return True  # keep the timer running
-
-    @staticmethod
-    def _format_speed(bps: float) -> str:
-        if bps >= 1024 * 1024:
-            return f"{bps / (1024 * 1024):.1f} MB/s"
-        if bps >= 1024:
-            return f"{bps / 1024:.1f} KB/s"
-        return f"{int(bps)} B/s"
-
-    def _set_dialog_heading(self, text: str) -> None:
-        """Set the dialog's primary heading on either base class."""
-        if _HAS_ALERT_DIALOG:
-            self.set_heading(text)
-        else:
-            self.set_title(text)
-    
-    def increment_file_count(self):
-        """Increment completed file counter"""
-        GLib.idle_add(self._increment_file_count_ui)
-    
-    def _increment_file_count_ui(self):
-        """Update file counter (must be called from main thread)"""
-        self.files_completed += 1
-        self.counter_label.set_text(f"{self.files_completed} of {self.total_files} files")
-        return False
-    
-    def set_future(self, future):
-        """Set the current operation future for cancellation"""
-        self._current_future = future
-        if future not in self._futures:
-            self._futures.append(future)
-    
-    def _format_time(self, seconds):
-        """Format time remaining for display"""
-        if seconds > 3600:
-            hours = int(seconds // 3600)
-            minutes = int((seconds % 3600) // 60)
-            return f"{hours}h {minutes}m remaining"
-        elif seconds > 60:
-            minutes = int(seconds // 60)
-            return f"{minutes}m remaining"
-        else:
-            return f"{int(seconds)}s remaining"
-    
-    def _format_size(self, size_bytes):
-        """Format file size for display"""
-        if size_bytes >= 1024 * 1024 * 1024:  # GB
-            return f"{size_bytes / (1024 * 1024 * 1024):.1f} GB"
-        elif size_bytes >= 1024 * 1024:  # MB
-            return f"{size_bytes / (1024 * 1024):.1f} MB"
-        elif size_bytes >= 1024:  # KB
-            return f"{size_bytes / 1024:.1f} KB"
-        else:
-            return f"{size_bytes} bytes"
-    
-    def show_completion(self, success=True, error_message=None):
-        """Show completion state"""
-        GLib.idle_add(self._show_completion_ui, success, error_message)
-    
-    def _show_completion_ui(self, success, error_message):
-        """Update UI to show completion state (idempotent - safe to call multiple times)"""
-        # Prevent multiple completion dialogs from being shown
-        if self._completion_shown:
-            return False
-        self._completion_shown = True
-        self._stop_render_timer()
-        
-        if success:
-            self._set_dialog_heading("Transfer Complete")
-            self.status_label.set_text("Transfer completed successfully")
-            self.file_label.set_text(f"Successfully transferred {self.files_completed} files")
-            self.progress_bar.set_fraction(1.0)
-            self.progress_bar.set_text("100%")
-        else:
-            self._set_dialog_heading("Transfer Failed")
-            self.status_label.set_text("Transfer failed")
-            if error_message:
-                self.file_label.set_text(f"Error: {error_message}")
-            else:
-                self.file_label.set_text("An error occurred during transfer")
-        
-        # Swap the body button's label from Cancel → Done. _on_action_button_clicked
-        # reads _completion_shown (set above) and takes the "just close" path.
-        try:
-            self.action_button.set_label("Done")
-            self.action_button.add_css_class("suggested-action")
-        except (AttributeError, RuntimeError, GLib.GError):
-            # Button may have been destroyed if the dialog is mid-close.
-            pass
-
-        return False
 
 
 # Icon size steps for the SFTP file manager. Both tuples are indexed by the
@@ -2498,635 +1996,6 @@ class AsyncSFTPManager(GObject.GObject):
 
 
 
-
-# ---------------------------------------------------------------------------
-# UI widgets
-
-
-class PathEntry(Gtk.Entry):
-    """Simple entry used for the editable pathbar."""
-
-    def __init__(self) -> None:
-        super().__init__()
-        # Don't set hexpand here - we'll set it explicitly in the toolbar
-        # self.set_hexpand(True)
-        self.set_placeholder_text("/remote/path")
-        # Remove minimum width constraint to allow full expansion
-        # self.set_size_request(200, -1)  # Commented out to allow full width
-
-
-class PaneControls(Gtk.Box):
-    def __init__(self) -> None:
-        super().__init__(orientation=Gtk.Orientation.HORIZONTAL, spacing=6)
-        self.set_valign(Gtk.Align.CENTER)
-        from sshpilot import icon_utils
-        self.back_button = icon_utils.new_button_from_icon_name("go-previous-symbolic")
-        self.up_button = icon_utils.new_button_from_icon_name("go-up-symbolic")
-        self.refresh_button = icon_utils.new_button_from_icon_name("view-refresh-symbolic")
-        self.new_folder_button = icon_utils.new_button_from_icon_name("folder-new-symbolic")
-        for widget in (
-            self.back_button,
-            self.up_button,
-            self.refresh_button,
-            self.new_folder_button,
-        ):
-            widget.set_valign(Gtk.Align.CENTER)
-        for widget in (self.back_button, self.up_button, self.refresh_button, self.new_folder_button):
-            widget.add_css_class("flat")
-        self.append(self.back_button)
-        self.append(self.up_button)
-        self.append(self.refresh_button)
-        self.append(self.new_folder_button)
-
-
-class PaneToolbar(Gtk.Box):
-    __gsignals__ = {
-        "view-changed": (GObject.SignalFlags.RUN_LAST, None, (str,)),
-        "show-hidden-toggled": (GObject.SignalFlags.RUN_LAST, None, (bool,)),
-        "zoom-changed": (GObject.SignalFlags.RUN_LAST, None, (int,)),
-    }
-
-    def __init__(self):
-        super().__init__(orientation=Gtk.Orientation.VERTICAL)
-
-        # Build a custom top bar: WindowHandle -> Box [ left | ENTRY (expands) | right ]
-        handle = Gtk.WindowHandle()                    # gives draggable area like a headerbar
-        bar = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, spacing=6)
-        handle.set_child(bar)
-
-        # Left side (compact)
-        self._pane_label = Gtk.Label()
-        self._pane_label.set_css_classes(["title"])
-        self.controls = PaneControls()
-        left = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, spacing=6)
-        left.set_margin_start(12)  # Add margin before Remote/Local labels
-        left.append(self._pane_label)
-        left.append(self.controls)
-        bar.append(left)
-
-        # Entry (fills all remaining space)
-        self.path_entry = PathEntry()
-        self.path_entry.set_hexpand(True)
-        self.path_entry.set_halign(Gtk.Align.FILL)
-        self.path_entry.set_width_chars(0)
-        self.path_entry.set_max_width_chars(0)
-        bar.append(self.path_entry)
-
-        # Right side (compact, flush-right)
-        right = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, spacing=6)
-        self._current_view = "list"
-
-        # Toggle for showing hidden files within this pane
-        self._show_hidden_button = Gtk.ToggleButton()
-        self._show_hidden_button.set_valign(Gtk.Align.CENTER)
-        self._show_hidden_button.add_css_class("flat")
-        from sshpilot import icon_utils
-        self._show_hidden_image = icon_utils.new_image_from_icon_name("view-conceal-symbolic")
-        self._show_hidden_button.set_child(self._show_hidden_image)
-        self._show_hidden_handler_id = self._show_hidden_button.connect(
-            "toggled", self._on_show_hidden_toggled
-        )
-        self._update_show_hidden_icon(False)
-        right.append(self._show_hidden_button)
-
-        self.sort_split_button = self._create_sort_split_button()
-        right.append(self.sort_split_button)
-        bar.append(right)
-
-        # Don't wrap in ToolbarView to avoid nested ToolbarView theme issues
-        # Instead, add toolbar styling via CSS class and append handle directly
-        handle.add_css_class('toolbar')
-        self.append(handle)
-
-    # Keep your factory
-    def _create_sort_split_button(self) -> Adw.SplitButton:
-        menu_model = Gio.Menu()
-
-        # Custom widget slot: the icon-size slider lives at the top of the
-        # dropdown. The "custom" attribute names a slot that PopoverMenu
-        # fills with whatever widget we add_child() under the same id.
-        size_section = Gio.Menu()
-        zoom_item = Gio.MenuItem.new("Icon Size", None)
-        zoom_item.set_attribute_value("custom", GLib.Variant.new_string("zoom_slider"))
-        size_section.append_item(zoom_item)
-        menu_model.append_section(None, size_section)
-
-        sort_section = Gio.Menu()
-        sort_section.append("Name", "pane.sort-by-name")
-        sort_section.append("Size", "pane.sort-by-size")
-        sort_section.append("Modified", "pane.sort-by-modified")
-        menu_model.append_section("Sort by", sort_section)
-        direction_section = Gio.Menu()
-        direction_section.append("Ascending", "pane.sort-direction-asc")
-        direction_section.append("Descending", "pane.sort-direction-desc")
-        menu_model.append_section("Order", direction_section)
-
-        # Build a PopoverMenu from the model so we can inject the slider widget
-        # into the named custom slot.
-        popover = Gtk.PopoverMenu.new_from_model(menu_model)
-        popover.add_child(self._build_zoom_slider_widget(), "zoom_slider")
-
-        split_button = Adw.SplitButton()
-        split_button.set_popover(popover)
-        split_button.set_tooltip_text("Toggle view mode")
-        # Tooltip text is parsed as Pango markup; escape the ampersand or
-        # use a plain word to avoid "entity did not end with a semicolon".
-        split_button.set_dropdown_tooltip("Adjust icon size and sort order")
-        from sshpilot import icon_utils
-        split_button.set_icon_name("view-list-symbolic")
-        split_button.connect("clicked", self._on_view_toggle_clicked)
-        return split_button
-
-    def _build_zoom_slider_widget(self) -> Gtk.Widget:
-        """Build the icon-size slider that sits inside the view-toggle dropdown."""
-        box = Gtk.Box(orientation=Gtk.Orientation.VERTICAL, spacing=2)
-        box.set_margin_top(6)
-        box.set_margin_bottom(2)
-        box.set_margin_start(10)
-        box.set_margin_end(10)
-        box.set_size_request(220, -1)
-
-        header = Gtk.Label(label="Icon Size", xalign=0)
-        header.add_css_class("caption-heading")
-        header.add_css_class("dim-label")
-        box.append(header)
-
-        scale = Gtk.Scale.new_with_range(
-            Gtk.Orientation.HORIZONTAL,
-            float(_MIN_ICON_LEVEL),
-            float(_MAX_ICON_LEVEL),
-            1.0,
-        )
-        scale.set_digits(0)
-        scale.set_draw_value(False)
-        scale.set_hexpand(True)
-        scale.set_round_digits(0)
-        for lvl in range(_MIN_ICON_LEVEL, _MAX_ICON_LEVEL + 1):
-            scale.add_mark(float(lvl), Gtk.PositionType.BOTTOM, None)
-        scale.set_value(float(_DEFAULT_ICON_LEVEL))
-        self._zoom_scale = scale
-        self._zoom_scale_handler_id = scale.connect(
-            "value-changed", self._on_zoom_scale_changed
-        )
-        box.append(scale)
-        return box
-
-    def _on_zoom_scale_changed(self, scale: Gtk.Scale) -> None:
-        level = int(round(scale.get_value()))
-        level = max(_MIN_ICON_LEVEL, min(_MAX_ICON_LEVEL, level))
-        self.emit("zoom-changed", level)
-
-    def set_zoom_level(self, level: int) -> None:
-        """Sync the slider to *level* without firing zoom-changed."""
-        scale = getattr(self, "_zoom_scale", None)
-        if scale is None:
-            return
-        clamped = float(max(_MIN_ICON_LEVEL, min(_MAX_ICON_LEVEL, level)))
-        if abs(scale.get_value() - clamped) < 0.5:
-            return
-        handler_id = getattr(self, "_zoom_scale_handler_id", None)
-        if handler_id is not None:
-            scale.handler_block(handler_id)
-        try:
-            scale.set_value(clamped)
-        finally:
-            if handler_id is not None:
-                scale.handler_unblock(handler_id)
-
-    # Example handler
-    def _on_view_toggle_clicked(self, *_):
-        from sshpilot import icon_utils
-        self._current_view = "grid" if self._current_view == "list" else "list"
-        icon_name = "view-grid-symbolic" if self._current_view == "grid" else "view-list-symbolic"
-        # Adw.SplitButton uses set_icon_name()
-        self.sort_split_button.set_icon_name(icon_name)
-        self.emit("view-changed", self._current_view)
-
-    def _on_show_hidden_toggled(self, button: Gtk.ToggleButton) -> None:
-        state = button.get_active()
-        self._update_show_hidden_icon(state)
-        self.emit("show-hidden-toggled", state)
-
-    def set_show_hidden_state(self, show_hidden: bool) -> None:
-        if not hasattr(self, "_show_hidden_button"):
-            return
-        current = self._show_hidden_button.get_active()
-        if current == show_hidden:
-            self._update_show_hidden_icon(show_hidden)
-            return
-        if self._show_hidden_handler_id is not None:
-            self._show_hidden_button.handler_block(self._show_hidden_handler_id)
-        try:
-            self._show_hidden_button.set_active(show_hidden)
-        finally:
-            if self._show_hidden_handler_id is not None:
-                self._show_hidden_button.handler_unblock(self._show_hidden_handler_id)
-        self._update_show_hidden_icon(show_hidden)
-
-    def _update_show_hidden_icon(self, show_hidden: bool) -> None:
-        from sshpilot import icon_utils
-        icon_name = "view-reveal-symbolic" if show_hidden else "view-conceal-symbolic"
-        tooltip = "Hide Hidden Files" if show_hidden else "Show Hidden Files"
-        icon_utils.set_icon_from_name(self._show_hidden_image, icon_name)
-        self._show_hidden_button.set_tooltip_text(tooltip)
-
-    def get_header_bar(self):
-        """Get the actual header bar for toolbar view."""
-        return None  # No longer using Adw.HeaderBar
-
-
-
-class PropertiesDialog(Adw.Window):
-    """Nautilus-style properties dialog using card-based design."""
-
-    def __init__(self, entry: "FileEntry", current_path: str, parent: Gtk.Window, sftp_manager: Optional["AsyncSFTPManager"] = None):
-        super().__init__()
-        self._entry = entry
-        self._current_path = current_path
-        self._parent_window = parent
-        self._sftp_manager = sftp_manager
-        self.set_title("Properties")
-        
-        # Set window properties
-        self.set_default_size(400, 500)
-        self.set_resizable(True)
-        self.set_modal(True)
-        self.set_transient_for(parent)
-        
-        # Position window relative to parent
-        if parent:
-            try:
-                # Get parent window position and size
-                parent_alloc = parent.get_allocation()
-                parent_width = parent_alloc.width
-                parent_height = parent_alloc.height
-                
-                # Center the dialog on the parent window
-                # For GTK4, we'll let the window manager handle positioning
-                # The modal and transient_for properties should handle this
-            except Exception:
-                # Fallback: let window manager handle positioning
-                pass
-
-        # Build the dialog content
-        self._build_dialog()
-
-    def _build_dialog(self) -> None:
-        """Build the Nautilus-style properties dialog content."""
-        # Create AdwToolbarView as the main content (proper Adw.Window structure)
-        toolbar_view = Adw.ToolbarView()
-        
-        # Create proper header bar for dragging
-        header_bar = Adw.HeaderBar()
-        header_bar.set_title_widget(Gtk.Label(label="Properties"))
-        
-        # Add header bar to toolbar view
-        toolbar_view.add_top_bar(header_bar)
-        
-        # Main content box
-        content = Gtk.Box(orientation=Gtk.Orientation.VERTICAL, spacing=16,
-                         margin_top=16, margin_bottom=16, margin_start=16, margin_end=16)
-        
-        # Header with icon and name
-        content.append(self._create_header_block())
-        
-        # Parent folder row
-        content.append(self._create_parent_folder_row())
-        
-        # Size row
-        content.append(self._create_size_row())
-        
-        # Modified and Created rows
-        content.append(self._create_modified_row())
-        content.append(self._create_created_row())
-        
-        # Permissions row
-        content.append(self._create_permissions_row())
-        
-        # Set content in toolbar view
-        toolbar_view.set_content(content)
-        
-        # Set the toolbar view as the window content
-        self.set_content(toolbar_view)
-
-
-    def _create_header_block(self) -> Gtk.Widget:
-        """Create the header block with icon, name, and summary."""
-        box = Gtk.Box(orientation=Gtk.Orientation.VERTICAL, spacing=8, halign=Gtk.Align.CENTER)
-        
-        # Icon — use the file-type-aware icon resolved from the entry's name.
-        from sshpilot import icon_utils
-        from .file_type_icons import get_icon_for_name
-        icon_name = get_icon_for_name(self._entry.name, self._entry.is_dir)
-        icon = icon_utils.new_image_from_icon_name(icon_name)
-        # Set a larger custom size instead of using predefined sizes
-        icon.set_pixel_size(64)
-        icon.add_css_class("icon-dropshadow")
-        icon.add_css_class("card")
-        box.append(icon)
-        
-        # Name (centered, bold)
-        name_label = Gtk.Label(label=self._entry.name)
-        name_label.add_css_class("title-3")
-        box.append(name_label)
-        
-        # Summary
-        summary_parts = []
-        if self._entry.is_dir:
-            if self._entry.item_count is not None:
-                summary_parts.append(f"{self._entry.item_count} item{'s' if self._entry.item_count != 1 else ''}")
-            else:
-                summary_parts.append("Folder")
-        else:
-            if self._entry.size:
-                summary_parts.append(_human_size(self._entry.size))
-        
-        # Add free space for local files
-        if not self._is_remote_file():
-            try:
-                path = os.path.join(self._current_path, self._entry.name)
-                if os.path.exists(path):
-                    stat = os.statvfs(path)
-                    free = stat.f_bavail * stat.f_frsize
-                    summary_parts.append(f"{_human_size(free)} Free")
-            except Exception:
-                pass
-        
-        summary_text = " — ".join(summary_parts) if summary_parts else ""
-        summary_label = Gtk.Label(label=summary_text)
-        summary_label.add_css_class("dim-label")
-        box.append(summary_label)
-        
-        return box
-
-    def _create_size_row(self) -> Gtk.Widget:
-        """Create the size row."""
-        if self._entry.is_dir:
-            if self._entry.item_count is not None:
-                size_text = f"{self._entry.item_count} item{'s' if self._entry.item_count != 1 else ''}"
-                # For local folders, start calculating actual size
-                if not self._is_remote_file():
-                    size_text += " (calculating size...)"
-                    self._start_folder_size_calculation()
-            else:
-                size_text = "—"
-        else:
-            size_text = _human_size(self._entry.size) if self._entry.size else "—"
-        
-        # Store reference to size row for updating
-        self._size_row = Adw.ActionRow(title="Size", subtitle=size_text)
-        self._size_row.add_css_class("card")
-        return self._size_row
-
-    def _create_parent_folder_row(self) -> Gtk.Widget:
-        """Create the parent folder row."""
-        parent_path = os.path.dirname(os.path.join(self._current_path, self._entry.name))
-        if not parent_path:
-            parent_path = "/"
-        
-        row = Adw.ActionRow(title="Parent Folder", subtitle=parent_path)
-        row.add_css_class("card")
-        
-        # Add folder open button for local files
-        if not self._is_remote_file():
-            from sshpilot import icon_utils
-            btn = icon_utils.new_button_from_icon_name("folder-open-symbolic")
-            btn.add_css_class("flat")
-            btn.connect("clicked", self._on_open_parent)
-            row.add_suffix(btn)
-            row.set_activatable_widget(btn)
-        
-        return row
-
-    def _create_modified_row(self) -> Gtk.Widget:
-        """Create the modified date row."""
-        modified_time = _human_time(self._entry.modified) if self._entry.modified else "—"
-        row = Adw.ActionRow(title="Modified", subtitle=modified_time)
-        row.add_css_class("card")
-        return row
-
-    def _create_created_row(self) -> Gtk.Widget:
-        """Create the created date row (if available)."""
-        # For remote files, we typically don't have creation time
-        if self._is_remote_file():
-            return Gtk.Box()  # Empty box widget
-        
-        # Try to get creation time for local files
-        try:
-            path = os.path.join(self._current_path, self._entry.name)
-            if os.path.exists(path):
-                stat_result = os.stat(path)
-                if hasattr(stat_result, 'st_birthtime'):  # macOS
-                    created_time = _human_time(stat_result.st_birthtime)
-                elif hasattr(stat_result, 'st_ctime'):  # Linux
-                    created_time = _human_time(stat_result.st_ctime)
-                else:
-                    return Gtk.Box()  # Empty box widget
-            else:
-                return Gtk.Box()  # Empty box widget
-        except Exception:
-            return Gtk.Box()  # Empty box widget
-        
-        row = Adw.ActionRow(title="Created", subtitle=created_time)
-        row.add_css_class("card")
-        return row
-
-    def _create_permissions_row(self) -> Gtk.Widget:
-        """Create the permissions row."""
-        perms_text = "—"
-        
-        # Get actual permissions for local files
-        if not self._is_remote_file():
-            try:
-                path = os.path.join(self._current_path, self._entry.name)
-                if os.path.exists(path):
-                    stat_result = os.stat(path)
-                    mode = stat_result.st_mode
-                    # Show both letter format and numeric format
-                    letter_format = _mode_to_str(mode)
-                    numeric_format = _mode_to_octal(mode)
-                    perms_text = f"{letter_format} ({numeric_format})"
-                else:
-                    perms_text = "—"
-            except Exception:
-                perms_text = "—"
-        else:
-            # For remote files, try to get permissions from SFTP asynchronously
-            # Start with a placeholder, will be updated when stat completes
-            perms_text = "Loading…"
-            
-            logger.debug(f"PropertiesDialog: Checking SFTP manager for remote file permissions")
-            logger.debug(f"PropertiesDialog: _sftp_manager={self._sftp_manager}")
-            
-            if self._sftp_manager:
-                has_sftp = hasattr(self._sftp_manager, '_sftp')
-                logger.debug(f"PropertiesDialog: hasattr(_sftp_manager, '_sftp')={has_sftp}")
-                if has_sftp:
-                    sftp_client = getattr(self._sftp_manager, '_sftp', None)
-                    logger.debug(f"PropertiesDialog: _sftp_manager._sftp={sftp_client}")
-                
-                if has_sftp and sftp_client:
-                    # Build the full remote path
-                    if self._current_path.endswith('/'):
-                        remote_path = self._current_path + self._entry.name
-                    else:
-                        remote_path = posixpath.join(self._current_path, self._entry.name)
-                    
-                    logger.debug(f"PropertiesDialog: Fetching permissions for remote path: {remote_path}")
-                    
-                    # Get file attributes from SFTP asynchronously
-                    def _get_attr():
-                        assert self._sftp_manager._sftp is not None
-                        logger.debug(f"PropertiesDialog: Background thread calling stat({remote_path})")
-                        attr = self._sftp_manager._sftp.stat(remote_path)
-                        logger.debug(f"PropertiesDialog: stat() returned: {attr}, st_mode={getattr(attr, 'st_mode', None)}")
-                        return attr
-                    
-                    def _update_permissions(future):
-                        try:
-                            logger.debug(f"PropertiesDialog: Future completed, getting result")
-                            attr = future.result()
-                            logger.debug(f"PropertiesDialog: Got attr: {attr}, has st_mode: {hasattr(attr, 'st_mode')}")
-                            if attr and hasattr(attr, 'st_mode'):
-                                mode = attr.st_mode
-                                # Show both letter format and numeric format
-                                letter_format = _mode_to_str(mode)
-                                numeric_format = _mode_to_octal(mode)
-                                new_text = f"{letter_format} ({numeric_format})"
-                                logger.debug(f"PropertiesDialog: Setting permissions to: {new_text}")
-                            else:
-                                # Fallback to simplified permissions
-                                if self._entry.is_dir:
-                                    new_text = "Create and Delete Files"
-                                else:
-                                    new_text = "Read and Write"
-                                logger.debug(f"PropertiesDialog: No st_mode, using fallback: {new_text}")
-                        except Exception as e:
-                            logger.debug(f"Failed to get remote file permissions: {e}", exc_info=True)
-                            # Fallback to simplified permissions if we can't get mode
-                            if self._entry.is_dir:
-                                new_text = "Create and Delete Files"
-                            else:
-                                new_text = "Read and Write"
-                        
-                        # Update the row subtitle on the main thread
-                        GLib.idle_add(lambda: self._update_permissions_row(new_text))
-                    
-                    # Submit stat operation to background thread
-                    logger.debug(f"PropertiesDialog: Submitting stat operation to background thread")
-                    future = self._sftp_manager._submit(_get_attr)
-                    future.add_done_callback(_update_permissions)
-                    logger.debug(f"PropertiesDialog: Future submitted, callback added")
-                else:
-                    logger.debug(f"PropertiesDialog: No SFTP client available")
-                    # No SFTP manager available, show simplified permissions
-                    if self._entry.is_dir:
-                        perms_text = "Create and Delete Files"
-                    else:
-                        perms_text = "Read and Write"
-            else:
-                logger.debug(f"PropertiesDialog: No SFTP manager available")
-                # No SFTP manager available, show simplified permissions
-                if self._entry.is_dir:
-                    perms_text = "Create and Delete Files"
-                else:
-                    perms_text = "Read and Write"
-        
-        row = Adw.ActionRow(title="Permissions", subtitle=perms_text)
-        row.add_css_class("card")
-        # Store reference to row for async updates
-        self._permissions_row = row
-        
-        return row
-    
-    def _update_permissions_row(self, text: str) -> None:
-        """Update the permissions row subtitle."""
-        if hasattr(self, '_permissions_row'):
-            self._permissions_row.set_subtitle(text)
-
-    def _is_remote_file(self) -> bool:
-        """Check if this is a remote file (from SFTP)."""
-        # Check if we have an SFTP manager - that's the most reliable indicator
-        if self._sftp_manager is not None:
-            logger.debug(f"PropertiesDialog: Detected remote file (has SFTP manager)")
-            return True
-        
-        # Fallback heuristic - in a real implementation, you'd pass connection info
-        is_remote = "://" in self._current_path or (self._current_path.startswith("/") and 
-                not os.path.exists(os.path.join(self._current_path, self._entry.name)))
-        logger.debug(f"PropertiesDialog: _is_remote_file()={is_remote}, path={self._current_path}, has_sftp_manager={self._sftp_manager is not None}")
-        return is_remote
-
-    def _on_open_parent(self, *_) -> None:
-        """Open parent directory in system file manager."""
-        try:
-            if not self._is_remote_file():
-                parent_dir = os.path.dirname(os.path.join(self._current_path, self._entry.name))
-                if os.path.exists(parent_dir):
-                    Gio.AppInfo.launch_default_for_uri(f"file://{parent_dir}", None)
-        except Exception:
-            pass
-
-    def _start_folder_size_calculation(self):
-        """Start calculating folder size in background thread."""
-        import threading
-        
-        folder_path = os.path.join(self._current_path, self._entry.name)
-        
-        # Create and start the background thread
-        thread = threading.Thread(target=self._calculate_folder_size, args=(folder_path,))
-        thread.daemon = True  # Allows main program to exit even if thread is running
-        thread.start()
-
-    def _calculate_folder_size(self, path):
-        """
-        Recursively calculates the size of a folder.
-        THIS RUNS ON A BACKGROUND THREAD.
-        """
-        total_size = 0
-        try:
-            for dirpath, dirnames, filenames in os.walk(path):
-                for f in filenames:
-                    fp = os.path.join(dirpath, f)
-                    # Skip if it is a symlink or file doesn't exist
-                    if not os.path.islink(fp):
-                        try:
-                            total_size += os.path.getsize(fp)
-                        except FileNotFoundError:
-                            # File might have been deleted while scanning
-                            pass
-                        except OSError:
-                            # Permissions error, etc.
-                            pass
-
-        except Exception:
-            total_size = -1  # Use a negative value to signal an error
-
-        # When done, schedule the UI update on the main GTK thread
-        GLib.idle_add(self._update_folder_size_ui, total_size)
-        
-    def _update_folder_size_ui(self, total_size):
-        """
-        Updates the size row with the final folder size.
-        THIS RUNS ON THE MAIN GTK THREAD.
-        """
-        if hasattr(self, '_size_row') and self._size_row:
-            if total_size >= 0:
-                human_readable_size = _human_size(total_size)
-                if self._entry.item_count is not None:
-                    size_text = f"{self._entry.item_count} item{'s' if self._entry.item_count != 1 else ''} ({human_readable_size})"
-                else:
-                    size_text = human_readable_size
-            else:
-                if self._entry.item_count is not None:
-                    size_text = f"{self._entry.item_count} item{'s' if self._entry.item_count != 1 else ''} (size unavailable)"
-                else:
-                    size_text = "Size unavailable"
-            
-            self._size_row.set_subtitle(size_text)
-            
-        # Returning GLib.SOURCE_REMOVE ensures this function only runs once
-        return GLib.SOURCE_REMOVE
 
 
 class FilePane(Gtk.Box):
